+++ _wotlk/Interface/Addons/RXPGuides/CHANGELOG.md
@@ -1,6 +1,10 @@
 # RestedXP Guides
 
-## [v4.6.30-1-g8650ed1](https://github.com/RestedXP/RXPGuides/tree/8650ed110f510726bbb974a415d155b49e436edc) (2024-07-23)
-[Full Changelog](https://github.com/RestedXP/RXPGuides/compare/v4.6.30...8650ed110f510726bbb974a415d155b49e436edc) [Previous Releases](https://github.com/RestedXP/RXPGuides/releases)
+## [v4.6.31-5-gc113012](https://github.com/RestedXP/RXPGuides/tree/c113012372a071df15ed1f3b19ad2ec22c5a561e) (2024-07-28)
+[Full Changelog](https://github.com/RestedXP/RXPGuides/compare/v4.6.31...c113012372a071df15ed1f3b19ad2ec22c5a561e) [Previous Releases](https://github.com/RestedXP/RXPGuides/releases)
 
-- Waypoint arrow fix  
+- Emerald wardens rune guide update  
+- SoD (A)leveling route updated with new book vendor  
+- Typo fixes  
+- #season typo fixes  
+- Small rune fixes